daemon: allow temp values to be floats for speed profiles (curves)
This is a general enhancement especially for liquid AIOs. Our internal speed control will benefit form this right away, but external profiles controllers, like some AIOs, would depend on the device driver and firmware implementations, if full support would really exist for float values.

+++ coolercontrold/src/repositories/liquidctl/liquidctl_repo.rs
@@ -701,7 +701,8 @@ struct FixedSpeedRequest {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 struct SpeedProfileRequest {
     channel: String,
-    profile: Vec<(u8, u8)>,
+    // INFO: There is a possibility that some liquidctl device drivers could cast temps to int
+    profile: Vec<(f64, u8)>,
     temperature_sensor: Option<u8>,
 }
